Rename TDB tests to different names
