+++ jena-db/jena-dboe-transaction/src/test/java/org/apache/jena/dboe/transaction/TS_Transactions.java
@@ -38,6 +38,4 @@ import org.junit.runners.Suite.SuiteClasses;
 
 })
 public class TS_Transactions {
-
 }
-
